split molecule class into base class + subclasses -> handle molecule specific stuff in subclasses, adjust thresholl for drawing a bit, deactivate OnCollisionEnter for performance sake and because it was never really called
