+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Catalyst.pc.unity
@@ -1768,12 +1768,12 @@ PrefabInstance:
     - target: {fileID: 1841423616163858762, guid: 8dec41cb1bd0f0a4e8beabcdd339d90f,
         type: 3}
       propertyPath: m_AnchorMax.x
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 1841423616163858762, guid: 8dec41cb1bd0f0a4e8beabcdd339d90f,
         type: 3}
       propertyPath: m_AnchorMax.y
-      value: 0
+      value: 0.7796314
       objectReference: {fileID: 0}
     - target: {fileID: 3549571858872179889, guid: 8dec41cb1bd0f0a4e8beabcdd339d90f,
         type: 3}
@@ -5563,14 +5563,38 @@ PrefabInstance:
       objectReference: {fileID: 514853967}
     - target: {fileID: 8226137981584974561, guid: 0eb3a977f5b2d5d47ba16b4106424a97,
         type: 3}
+      propertyPath: oMoleculePrefab
+      value: 
+      objectReference: {fileID: 684029413386375373, guid: 62d7f2b7dfba5a241bb94438f2bc8395,
+        type: 3}
+    - target: {fileID: 8226137981584974561, guid: 0eb3a977f5b2d5d47ba16b4106424a97,
+        type: 3}
       propertyPath: variantDropdown
       value: 
       objectReference: {fileID: 1526018783}
     - target: {fileID: 8226137981584974561, guid: 0eb3a977f5b2d5d47ba16b4106424a97,
         type: 3}
+      propertyPath: coMoleculePrefab
+      value: 
+      objectReference: {fileID: 2372393903890976028, guid: ef61a79bd7a3f894d85c0ec943d4138d,
+        type: 3}
+    - target: {fileID: 8226137981584974561, guid: 0eb3a977f5b2d5d47ba16b4106424a97,
+        type: 3}
+      propertyPath: o2MoleculePrefab
+      value: 
+      objectReference: {fileID: 8374867824179807448, guid: 78b5e44d54114f24280652f2443008d9,
+        type: 3}
+    - target: {fileID: 8226137981584974561, guid: 0eb3a977f5b2d5d47ba16b4106424a97,
+        type: 3}
+      propertyPath: co2MoleculePrefab
+      value: 
+      objectReference: {fileID: 3049424487111058320, guid: 01362201552f40047bbce31c9e8f4b6b,
+        type: 3}
+    - target: {fileID: 8226137981584974561, guid: 0eb3a977f5b2d5d47ba16b4106424a97,
+        type: 3}
       propertyPath: cobaltMoleculePrefab
       value: 
-      objectReference: {fileID: 9201737708395207697, guid: 4f078f09cfce2274e9ae8604e8a2199b,
+      objectReference: {fileID: 6807192493002849540, guid: 4f078f09cfce2274e9ae8604e8a2199b,
         type: 3}
     - target: {fileID: 8226137981584974561, guid: 0eb3a977f5b2d5d47ba16b4106424a97,
         type: 3}
