+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Prefabs/CO2Molecule.prefab
@@ -562,7 +562,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 9201737708395207698}
-  - component: {fileID: 9201737708395207697}
+  - component: {fileID: 3049424487111058320}
   - component: {fileID: 69340141852692580}
   - component: {fileID: 2293727733925459083}
   m_Layer: 0
@@ -593,7 +593,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &9201737708395207697
+--- !u!114 &3049424487111058320
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -602,11 +602,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 9201737708395207699}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a59b96b1b42e42e98d57da113b3ce83c, type: 3}
+  m_Script: {fileID: 11500000, guid: 2e3d52d59f284dfb9cc720581c4dd1ba, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   type: 1
-  canConnectToList: 
   collider: {fileID: 69340141852692580}
   movementSpeed: 1
   timeToMove: 3
@@ -640,12 +639,6 @@ MonoBehaviour:
       m_PersistentCalls:
         m_Calls: []
   _state: 0
-  _temperatureStageValues: fa000000130100002c010000450100005e0100007701000090010000a9010000c2010000
-  _partialPressureValues:
-  - 0.01
-  - 0.02
-  - 0.04
-  - 0.2
 --- !u!136 &69340141852692580
 CapsuleCollider:
   m_ObjectHideFlags: 0
