+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Prefabs/COMolecule.prefab
@@ -325,7 +325,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 9201737708395207698}
-  - component: {fileID: 9201737708395207697}
+  - component: {fileID: 2372393903890976028}
   - component: {fileID: 3990006234335495563}
   - component: {fileID: 8644056029575331758}
   m_Layer: 0
@@ -353,7 +353,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &9201737708395207697
+--- !u!114 &2372393903890976028
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -362,11 +362,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 9201737708395207699}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a59b96b1b42e42e98d57da113b3ce83c, type: 3}
+  m_Script: {fileID: 11500000, guid: 901ac7a8b1a34a18adc7b23d70b5ccc7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   type: 0
-  canConnectToList: 0200000004000000
   collider: {fileID: 3990006234335495563}
   movementSpeed: 1
   timeToMove: 3
@@ -400,12 +399,6 @@ MonoBehaviour:
       m_PersistentCalls:
         m_Calls: []
   _state: 0
-  _temperatureStageValues: fa000000130100002c010000450100005e0100007701000090010000a9010000c2010000
-  _partialPressureValues:
-  - 0.01
-  - 0.02
-  - 0.04
-  - 0.2
 --- !u!136 &3990006234335495563
 CapsuleCollider:
   m_ObjectHideFlags: 0
