+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Prefabs/CatalystSurface.prefab
@@ -613,8 +613,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6751aa3360b24ab9ba06baab793ca7d7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  platinumMoleculePrefab: {fileID: 9201737708395207697, guid: d21d88ccf0215ea48aa2502b6c828982,
+  platinumMoleculePrefab: {fileID: 2647887616471056261, guid: d21d88ccf0215ea48aa2502b6c828982,
     type: 3}
-  coMoleculePrefab: {fileID: 9201737708395207697, guid: ef61a79bd7a3f894d85c0ec943d4138d,
+  coMoleculePrefab: {fileID: 2372393903890976028, guid: ef61a79bd7a3f894d85c0ec943d4138d,
     type: 3}
   surfaceLayerParent: {fileID: 7909708731039112216}
