+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Prefabs/CobaltMolecule.prefab
@@ -102,7 +102,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 9201737708395207698}
-  - component: {fileID: 9201737708395207697}
+  - component: {fileID: 6807192493002849540}
   - component: {fileID: 2216740001849344916}
   - component: {fileID: 2367873597819704465}
   m_Layer: 0
@@ -127,7 +127,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &9201737708395207697
+--- !u!114 &6807192493002849540
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -136,7 +136,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 9201737708395207699}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a59b96b1b42e42e98d57da113b3ce83c, type: 3}
+  m_Script: {fileID: 11500000, guid: bdd8d6ee31f949eaa9ff43141daf6362, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   type: 5
