+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Prefabs/O2Molecule.prefab
@@ -246,7 +246,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 9201737708395207698}
-  - component: {fileID: 9201737708395207697}
+  - component: {fileID: 8374867824179807448}
   - component: {fileID: 5281235216544359750}
   - component: {fileID: 4168725840538543636}
   m_Layer: 0
@@ -273,7 +273,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &9201737708395207697
+--- !u!114 &8374867824179807448
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -282,11 +282,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 9201737708395207699}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a59b96b1b42e42e98d57da113b3ce83c, type: 3}
+  m_Script: {fileID: 11500000, guid: a5c399bbd5fc40c4843faee570e5e9f5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   type: 3
-  canConnectToList: 04000000
   collider: {fileID: 5281235216544359750}
   movementSpeed: 1
   timeToMove: 3
@@ -320,12 +319,6 @@ MonoBehaviour:
       m_PersistentCalls:
         m_Calls: []
   _state: 0
-  _temperatureStageValues: fa000000130100002c010000450100005e0100007701000090010000a9010000c2010000
-  _partialPressureValues:
-  - 0.01
-  - 0.02
-  - 0.04
-  - 0.2
 --- !u!136 &5281235216544359750
 CapsuleCollider:
   m_ObjectHideFlags: 0
