+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Prefabs/OMolecule.prefab
@@ -102,7 +102,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 9201737708395207698}
-  - component: {fileID: 9201737708395207697}
+  - component: {fileID: 684029413386375373}
   - component: {fileID: 6053066872662406567}
   - component: {fileID: 4612165014388229691}
   m_Layer: 0
@@ -127,7 +127,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &9201737708395207697
+--- !u!114 &684029413386375373
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -136,11 +136,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 9201737708395207699}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a59b96b1b42e42e98d57da113b3ce83c, type: 3}
+  m_Script: {fileID: 11500000, guid: 531482ff3096459b89630054ddaf87c4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   type: 2
-  canConnectToList: 00000000
   collider: {fileID: 6053066872662406567}
   movementSpeed: 1
   timeToMove: 3
@@ -174,12 +173,6 @@ MonoBehaviour:
       m_PersistentCalls:
         m_Calls: []
   _state: 0
-  _temperatureStageValues: fa000000130100002c010000450100005e0100007701000090010000a9010000c2010000
-  _partialPressureValues:
-  - 0.01
-  - 0.02
-  - 0.04
-  - 0.2
 --- !u!135 &6053066872662406567
 SphereCollider:
   m_ObjectHideFlags: 0
