+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Prefabs/PtMolecule.prefab
@@ -102,7 +102,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 9201737708395207698}
-  - component: {fileID: 9201737708395207697}
+  - component: {fileID: 2647887616471056261}
   - component: {fileID: 2216740001849344916}
   - component: {fileID: 2367873597819704465}
   m_Layer: 0
@@ -127,7 +127,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &9201737708395207697
+--- !u!114 &2647887616471056261
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -136,15 +136,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 9201737708395207699}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a59b96b1b42e42e98d57da113b3ce83c, type: 3}
+  m_Script: {fileID: 11500000, guid: 066eeafe6eb644578223401e5a69328f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   type: 4
-  canConnectToList: 00000000
   collider: {fileID: 2216740001849344916}
   movementSpeed: 0
   timeToMove: 3
-  timeUntilNextDesorb: 3
   temperature:
     assessmentName: 
     isDynamic: 0
@@ -174,12 +172,6 @@ MonoBehaviour:
       m_PersistentCalls:
         m_Calls: []
   _state: 0
-  _temperatureStageValues: fa000000130100002c010000450100005e0100007701000090010000a9010000c2010000
-  _partialPressureValues:
-  - 0.01
-  - 0.02
-  - 0.04
-  - 0.2
 --- !u!136 &2216740001849344916
 CapsuleCollider:
   m_ObjectHideFlags: 0
