+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Scripts/CatalystSurface.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Maroon.scenes.experiments.Catalyst.Scripts.Molecules;
 using UnityEngine;
 
 namespace Maroon.scenes.experiments.Catalyst.Scripts
