+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Scripts/Molecules/CO2Molecule.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace Maroon.scenes.experiments.Catalyst.Scripts.Molecules
+{
+    public class CO2Molecule : Molecule
+    {
+
+    }
+}
