+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Scripts/Molecules/CobaltMolecule.cs
@@ -1,0 +1,7 @@
+﻿namespace Maroon.scenes.experiments.Catalyst.Scripts.Molecules
+{
+    public class CobaltMolecule : Molecule
+    {
+        
+    }
+}
